pyproject.toml
[tool.poetry]
name = "nonebot-plugin-WWwiki"
version = "0.0.1"
description = "发送kurobbs的数据"
authors = ["shiyingyingjiang <2798134864@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_WWwiki"}]
homepage = "https://github.com/shi-yingyingjiang/nonebot-plugin-WWwiki"
repository = "https://github.com/shi-yingyingjiang/nonebot-plugin-WWwiki"


[tool.poetry.dependencies]
python = ">=3.8"
nonebot2 = ">=2.2.1"
nonebot-adapter-onebot = ">=2.2.3"
jinja2 = ">=3.1.2"
httpx = ">=0.27.0"
pandas = ">=2.0.3"
pydantic = ">=1.10.11"
playwright = ">=1.33.0"
Pilow = ">=9.4.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"